requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
pydantic==2.7.4
redis==5.0.4
httpx==0.27.0
google-generativeai==0.7.1
python-multipart==0.0.9
